Added create function to Base.js.
Also added schema to gradeLevel class.
